extract plan benefits into separate module
